+++ CHANGES.md
@@ -1,4 +1,6 @@
 #5.0.0.beta1 20??-??-??
+ - 2015-06-11 Added possibility to define read only SysConfig settings.
+ - 2015-06-11 Added possibility to disable a Date/Time field (created with Layout object BuildDateSelection()).
  - 2015-06-10 Added possibility to show tickets from subqueues in AgenTicketZoom.
    Default on initial view is configurable (show or hide subqueue tickets).
  - 2015-06-08 Updated 7 day stats dashboard widget to use d3 instead of flot and removed flot completely.